+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - 2017-09-24
+## [Unreleased]
+
+## [v1.3.0] - 2017-10-04
 ### Added
 - Support for placeholders on PyStrings and TableNodes
 - Placeholders chaining: use placeholders inside a placeholder
@@ -25,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Behat --tags filter now works correctly on variant tagged scenarios
 
-[Unreleased]: https://github.com/ciandt-dev/behat-placeholders-extension/compare/v1.2.1...HEAD
+[Unreleased]: https://github.com/ciandt-dev/behat-placeholders-extension/compare/v1.3.0...HEAD
+[v1.3.0]: https://github.com/ciandt-dev/behat-placeholders-extension/compare/v1.2.1...v1.3.0
 [v1.2.1]: https://github.com/ciandt-dev/behat-placeholders-extension/compare/v1.2.0...v1.2.1
 [v1.2.0]: https://github.com/ciandt-dev/behat-placeholders-extension/compare/v1.1.1...v1.2.0